Add example modulo instruction to Spike ISA simulator
The encoding is generated by the riscv-opcodes repository. We then
declare the instruction to Spike, so it understands the instruction
existing, allowing it to print the instruction correctly, rather than
just printin `unknown`.

We also add a test for the `mod` instruction that is identical to the
`rem` instruction, because I have said they will have the same semantics.